Enable PCI device enumeration on Compute Module 4
 * The PCI offset may be a 64-bit number like 0x600000000, which needs
   to fit in off_t (for mmap) and board_struct.mem_base
 * off_t is only 64-bits with -D_FILE_OFFSET_BITS=64
   (which is not the default in 2021!!!)
 * This requires the offset to be printed as a uint64_t
 * mmap was not checked for failure

After this change, I can detect a 5i24 plugged into the 1x slot of
a Raspberry Pi Compute Module 4 IO board with 2GB RAM compute module.
That's the only operation I tried.

+++ boards.h
@@ -60,7 +60,7 @@ struct board_struct {
     struct pci_dev *dev;
     void *base;
     int len;
-    u32 mem_base;
+    int64_t mem_base;
 #ifdef _WIN32
     tagPhysStruct_t mem_handle;
 #endif
